Updated brewhouse tests to use entityTestFactory
